requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
audioop-lts==0.2.1
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.9
distro==1.9.0
flaredantic==0.1.4
Flask==3.1.1
Flask-SocketIO==5.5.1
gevent==25.5.1
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.45
GPUtil==1.4.0
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
narwhals==2.5.0
numpy==2.3.3
openai==1.60.2
packaging==25.0
pandas==2.3.2
pillow==11.2.1
plotly==6.3.0
pluggy==1.6.0
protobuf==6.32.1
psutil==6.1.1
pyarrow==21.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.2
pypiwin32==223
pyshorteners==1.0.1
pytesseract==0.3.13
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pyttsx3==2.98
pytz==2025.2
pywin32==308
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SpeechRecognition==3.14.1
speedtest-cli==2.1.3
SQLAlchemy==2.0.43
standard-aifc==3.13.0
standard-chunk==3.13.0
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.1.1
zope.interface==7.2
